ré-organise le code des graphes indicateur
