+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/IndicateurEnfant.tsx
@@ -1,13 +1,13 @@
-import {useIndicateurACompleter} from './useIndicateurDefinitions';
+import classNames from 'classnames';
 import {TIndicateurReferentielDefinition} from './types';
 import {IndicateurValuesTabs} from './IndicateurValuesTabs';
 import {Badge} from 'ui/shared/Badge';
-import classNames from 'classnames';
+import {BadgeACompleter} from 'ui/shared/Badge/BadgeACompleter';
 import {useToggle} from 'ui/shared/useToggle';
 import {ExpandToggle} from 'ui/icons/ExpandToggle';
-import {BadgeACompleter} from '../../../../ui/shared/Badge/BadgeACompleter';
+import {useIndicateurACompleter} from './useIndicateurDefinitions';
 import {ActionsLieesCards} from '../PlansActions/FicheAction/FicheActionForm/ActionsLiees';
-import IndicateurChart from './IndicateurChart';
+import IndicateurChart from './charts/IndicateurChart';
 import {FichesActionLiees} from './FichesActionLiees';
 
 export const IndicateurEnfant = ({
