+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/charts/IndicateurChart.stories.tsx
