+++ app.territoiresentransitions.react/src/app/pages/collectivite/Indicateurs/charts/types.ts
@@ -1,0 +1,13 @@
+import {TIndicateurDefinition} from '../types';
+import {TIndicateurValeur} from '../useIndicateurValeurs';
+
+export type TIndicateurChartProps = {
+  definition: TIndicateurDefinition;
+  /** affichage par défaut (dans la grille) ou zoomé (téléchargeable) */
+  variant?: 'default' | 'zoomed';
+  className?: string;
+};
+
+export type TIndicateurChartBaseProps = TIndicateurChartProps & {
+  valeurs: TIndicateurValeur[];
+};
